src/app/cashplay/item/search-items/search-items.component.spec.ts
import { async, ComponentFixture, fakeAsync, TestBed } from '@angular/core/testing';

import { SearchItemsComponent } from './search-items.component';
import { MdButtonModule, MdInputModule } from '@angular/material';
import { SharedModule } from '../../../shared/shared.module';
import { By } from '@angular/platform-browser';
import { APP_CONFIG, CASHPLAY_CONFIG } from '../../../AppConfig';
import { ItemService } from '../item.service';
import { FormsModule } from '@angular/forms';
import { TestResource } from '../../../test-helper';
import { Observable } from 'rxjs/Observable';
import 'rxjs/Observable/of';
import { IPaginatedResource } from '../../../api.service';

describe('SearchItemsComponent', () => {
  let component: SearchItemsComponent;
  let fixture: ComponentFixture<SearchItemsComponent>;
  let itemServiceStub: ItemService = new ItemService(null);
  const items: TestResource[] = [
    {id: 1, name: 'foo'},
    {id: 2, name: 'bar'}
  ];

  beforeEach(async(() => {
    spyOn(itemServiceStub, 'search')
      .and.callThrough()
      .and.returnValue(Observable.of<IPaginatedResource<TestResource[]>>());

    TestBed.configureTestingModule({
      declarations: [
        SearchItemsComponent
      ],
      providers: [
        {provide: ItemService, useValue: itemServiceStub},
        {provide: APP_CONFIG, useValue: CASHPLAY_CONFIG},
      ],
      imports: [
        FormsModule,
        SharedModule,
        MdInputModule,
        MdButtonModule,
      ]
    })
      .compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(SearchItemsComponent);
    itemServiceStub = TestBed.get(ItemService);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should be created', () => {
    expect(component).toBeTruthy();
  });
  it('should disable search button if q is empty and enable if not', () => {
    const searchBtn = fixture.debugElement.query(By.css('.art-search-button'));
    expect(searchBtn.nativeElement.disabled).toBeTruthy();
    component.q = 'some text';
    fixture.detectChanges();
    expect(searchBtn.nativeElement.disabled).toBeFalsy();
  });
  it('should not display art-resource-list if dataLength is zero', () => {
    component.dataLength = 0;
    fixture.detectChanges();
    expect(fixture.debugElement.query(By.css('art-resource-list.hide'))).toBeTruthy();
    component.dataLength = 1;
    fixture.detectChanges();
    expect(fixture.debugElement.query(By.css('art-resource-list.hide'))).toBeFalsy();
  });
  // FIXME fix this test
  it('should fill foundItems with calling search in service', fakeAsync(() => {
    expect(true).toEqual(true);
    // let f;
    // component.foundItems.subscribe(i => f = i);
    // component.ngOnInit();
    // // tick();
    // fixture.detectChanges();
    // component.searchTerm.next('foo');
    // fixture.detectChanges();
    // // tick();
    // fixture.detectChanges();
    // discardPeriodicTasks();
    // // expect(f.length).toEqual(2);
    // expect(fixture.debugElement.queryAll(By.css('md-row')).length).toEqual(2);
  }));
});
